Made new enemies that should be affected by getting hit by certain bullet colors
I made 4 new enemies, Red, Blue, Green, Yellow that are weak to a
certain bullet color. If they get hit by a color that is not their
weakness, they take half damage. I did this by making a new GetHit
function in each of their scripts that also takes in a color as a
parameter. Needs different bullet colors in order to be tested